+++ src/ai_watermark_toolkit/__init__.py
@@ -1,4 +1,4 @@
 __all__ = [
     "pipeline",
 ]
-__version__ = "2.3.0"
+__version__ = "2.3.1"